web_ui/requirements.txt
# Web UI specific dependencies
Flask==3.1.0
Flask-CORS==5.0.0
aiohttp==3.12.13
gunicorn==23.0.0
requests==2.32.3
Werkzeug==3.1.3

# These are already in the main requirements.txt but listed here for clarity
# when running the web UI independently
python-dotenv==1.1.0
